change: Move MediaUploadApi to the correct package
