+++ app/src/main/java/app/pachli/components/compose/MediaUploader.kt
@@ -31,8 +31,8 @@ import app.pachli.R
 import app.pachli.components.compose.ComposeActivity.QueuedMedia
 import app.pachli.core.common.string.randomAlphanumericString
 import app.pachli.core.data.model.InstanceInfo
-import app.pachli.core.mastodon.model.MediaUploadApi
 import app.pachli.core.network.extensions.getServerErrorMessage
+import app.pachli.core.network.model.MediaUploadApi
 import app.pachli.network.ProgressRequestBody
 import app.pachli.util.MEDIA_SIZE_UNKNOWN
 import app.pachli.util.getImageSquarePixels
